backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
pytest==8.4.1
sqlalchemy==2.0.42
psycopg[binary]==3.2.9
python-multipart==0.0.20
minio==7.2.16
httpx==0.28.1
pypdf==5.9.0
slowapi==0.1.9
alembic==1.16.1
structlog==25.4.0
pytest-cov==6.2.1
prometheus-fastapi-instrumentator==7.0.2
ruff==0.11.13
ruff==0.11.13
